finished api, almost done with viewroute redirection
