+++ seeds/index.js
@@ -4,12 +4,12 @@ const { hash } = require('bcrypt');
 const seedDatabase = async () => {
   await Teacher.bulkCreate([
     {
-      name: 'Teacher One', // Changed from username to name
+      name: 'Teacher One', 
       email: 'teacher1@example.com',
       password: await hash('examplepassword', 10)
     },
     {
-      name: 'Teacher Two', // Changed from username to name
+      name: 'Teacher Two', 
       email: 'teacher2@example.com',
       password: await hash('examplepassword', 10)
     }
@@ -27,9 +27,11 @@ const seedDatabase = async () => {
       teacher_id: 2
     }
   ]);
+  
 
   console.log('Database seeded!');
   process.exit(0);
 };
+
 
 seedDatabase();
